BF: Missing import on clone
Recent refactoring of `clone` (c57de3885cdbad488803241b63ab8d6b6e0738ec)
moved functions to `clone_utils.py` and forgot to import all of them.
Similarly, the introduction of `clone_ria.py` came with the same issue,
aa functions are moved. This led to breaking extension (datalad-next)
importing from clone.

This patch makes all those functions importable from clone again.

Note, that after this is released, datalad-next can drop its clone_ria
entirely, since it's part of core.

Closes datalad-next#111
